+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Add AppenderDynatrace
 * Add AppenderPool (thx @jimbrig)
+* add AppenderAWSCloudWatchLog for logging to [AWS CloudWatch](https://docs.aws.amazon.com/AmazonCloudWatch/latest/logs/WhatIsCloudWatchLogs.html) (#9) (thx @DyfanJones)
 
 
 # lgrExtra 0.0.9